% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist_to_border}
\alias{dist_to_border}
\title{Calculate the distance to the border of a survey region}
\usage{
dist_to_border(east, north, survey_region)
}
\arguments{
\item{east, }{@param north coordinates of the point}

\item{survey_region}{name of the survey region}
}
\value{
distance to the border of the survey region
}
\description{
Calculate the distance to the border of a survey region
}
