% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs.R
\name{fig1A}
\alias{fig1A}
\title{Figure 1A}
\usage{
fig1A()
}
\value{
A ggplot object.
}
\description{
Original and transformed coordinate axes.
The y′ axis is aligned with the border (dashed line) to the neighbouring region (shaded area)
while the x′ axis passes through the mean observed coordinate (filled point).
In the new coordinate system, observed coordinates (circles) are represented by their distance
d = x′ to the border (represented by dotted lines). The circle shows a 90% contour line of the
home region with center μ (cross).
}
\examples{
fig1A()
}
