% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs.R
\name{fig1B}
\alias{fig1B}
\title{Figure 1B}
\usage{
fig1B()
}
\value{
A ggplot object.
}
\description{
MLE of delta as a function of observed average distance to border d in units of sigma,
for selected ratios of lambda and k, where n is Poisson(lambda).
The absolute value |delta| correspond to the estimated distance of the activity center to the
border while the sign denotes its side with negative values in the neighbouring region
}
\examples{
fig1B()
}
