% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs.R
\name{fig2}
\alias{fig2}
\title{Figure 2}
\usage{
fig2()
}
\value{
A ggplot object.
}
\description{
The four regions of Swedish national monitoring of the brown bear. Dots represent average scat sample
locations of unique individuals encountered in each of the regions most recent survey.
}
\examples{
fig2()
}
