% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs.R
\name{fig3A}
\alias{fig3A}
\title{Figure 3A}
\usage{
fig3A()
}
\value{
A ggplot object.
}
\description{
Number of individuals with a given number of traces recorded (bars). Filled circles represent a
zero-truncated Negative Binomial fit and empty a zero-truncated Poisson.
The x-axis is truncated, excluding a small number of individuals with k > 20
}
\examples{
fig3A()
}
