% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs.R
\name{fig3B}
\alias{fig3B}
\title{Figure 3B}
\usage{
fig3B()
}
\value{
A ggplot object.
}
\description{
Histograms of observed individual coordinate sample standard deviations (grey). Dashed
lines show level of truncation. Outlined (black lines) are corresponding histograms
based on simulated coordinates under model assumptions with σ = spool.
The x-axis is truncated, excluding a small number of individuals with s > 50000
}
\examples{
fig3B()
}
