% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs.R
\name{fig4A}
\alias{fig4A}
\title{Figure 4A}
\usage{
fig4A()
}
\value{
A ggplot object.
}
\description{
Illustration of the relocation process in a high density area along the border between Region C and D.
Arrows, colored according to survey, start at the observed sample mean coordinate and point at the
estimated activity center.
}
\examples{
fig4A()
}
