% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs.R
\name{fig4B}
\alias{fig4B}
\title{Figure 4B}
\usage{
fig4B()
}
\value{
A ggplot object.
}
\description{
Number of individuals placed outside the survey region in relation to estimated home range size σ and
sampling distribution. Dots show estimated numbers based on the home range sizes reported in Table 1, lines show
fitted proportional relations (straight lines with zero intercept)
}
\examples{
fig4B()
}
