% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs.R
\name{fig5}
\alias{fig5}
\title{Figure 5}
\usage{
fig5()
}
\value{
A ggplot object.
}
\description{
Kernel density estimates of simulated values for delta. Means of simulated values are represented by vertical
dashed lines. Negative values corresponds to individuals in the neighbouring region.
}
\examples{
fig5()
}
