% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs.R
\name{fig6}
\alias{fig6}
\title{Figure 6}
\usage{
fig6()
}
\value{
A ggplot object.
}
\description{
Probability of a discovered individual being placed within the survey region given the distance|delta| of its
activity center to the border. Negative delta corresponds to individuals located in the neighbouring region
}
\examples{
fig6()
}
