% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit_center_Po}
\alias{fit_center_Po}
\title{Numerically optimize the log-likelihood of the Poisson model}
\usage{
fit_center_Po(east, north, k, sigma, region_name, lambda)
}
\arguments{
\item{east, }{@param north average sample coordinates}

\item{k}{number of samples}

\item{sigma}{standard deviation of home range distribution}

\item{region_name}{name of the region in which the samples were collected (as in survey_regions table)}

\item{lambda}{mean of the Poisson distribution}
}
\value{
: estimated center of the home range and proportion of the region occupied by the home range
}
\description{
Numerically optimize the log-likelihood of the Poisson model
}
