% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit_ztNB}
\alias{fit_ztNB}
\title{Numerically optimize the log-likelihood of the zero-truncated Negative Binomial distribution}
\usage{
fit_ztNB(n)
}
\arguments{
\item{n}{vector of counts}
}
\value{
estimated parameters of the zero-truncated Negative Binomial distribution
}
\description{
Numerically optimize the log-likelihood of the zero-truncated Negative Binomial distribution
}
