% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit_ztPo}
\alias{fit_ztPo}
\title{Numerically optimize the log-likelihood of the zero-truncated Poisson distribution}
\usage{
fit_ztPo(n)
}
\arguments{
\item{n}{vector of counts}
}
\value{
estimated parameter of the zero-truncated Poisson distribution
}
\description{
Numerically optimize the log-likelihood of the zero-truncated Poisson distribution
}
