% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in_region}
\alias{in_region}
\title{Check if a point is in a survey region}
\usage{
in_region(east, north, region)
}
\arguments{
\item{east, }{@param north coordinates of the point}

\item{region}{name of the survey region}
}
\value{
TRUE if the point is in the survey region, FALSE otherwise
}
\description{
Check if a point is in a survey region
}
