% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{LpolycubNB}
\alias{LpolycubNB}
\title{Calculate the log-likelihood of the Negative Binomial model}
\usage{
LpolycubNB(mu, x, y, k, sigma, region, lambda, theta)
}
\description{
Calculate the log-likelihood of the Negative Binomial model
}
