% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{LpolycubPo}
\alias{LpolycubPo}
\title{Calculate the log-likelihood of the Poisson model}
\usage{
LpolycubPo(mu, x, y, k, sigma, region, lambda)
}
\description{
Calculate the log-likelihood of the Poisson model
}
