% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_data}
\alias{raw_data}
\title{Raw coordinates of bears in the most recent survey of each survey region}
\format{
A data frame with 9479 rows and 5 variables:
\describe{
  \item{id}{ID of the bear from Rovbase 3.0}
  \item{survey_region}{Region where the bear was observed}
  \item{date}{Date of collection of the scat sample}
  \item{east}{Easting of the scat sample (SWEREF 99 TM)}
  \item{north}{Northing of the scat sample (SWEREF 99 TM)}
}
}
\usage{
raw_data
}
\description{
Raw coordinates of bears in the most recent survey of each survey region
}
\keyword{datasets}
