% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{region}
\alias{region}
\title{Find the survey region of a point.
 If the point is not in a survey region, it is assigned to either Norway (west of survey region) or
 Bothnian bay (east of survey region)}
\usage{
region(east, north)
}
\arguments{
\item{east, }{@param north coordinates of the point}
}
\value{
name of the survey region
}
\description{
Find the survey region of a point.
 If the point is not in a survey region, it is assigned to either Norway (west of survey region) or
 Bothnian bay (east of survey region)
}
