% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sigma_table}
\alias{sigma_table}
\title{Calculate three versions of sigma by sex}
\usage{
sigma_table(data, trim = 20000)
}
\arguments{
\item{trim}{threshold for the trimmed mean}
}
\value{
data frame with three versions of sigma (ac_sd) and the method used to calculate it (sd_method)
}
\description{
Calculate three versions of sigma by sex
}
