% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_regions}
\alias{survey_regions}
\title{Polygons of survey regions}
\format{
An sf object with four polygons corresponding to survey region borders
\describe{
  \item{survey_region}{Name of the survey region}
  \item{geometry}{Polygon of the survey region (SWEREF 99 TM)}
}
}
\usage{
survey_regions
}
\description{
Polygons of survey regions
}
\keyword{datasets}
