% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{tab1}
\alias{tab1}
\title{Table 1
Estimates of brown bear home range size sigma in meters.}
\usage{
tab1()
}
\value{
A data frame.
}
\description{
Table 1
Estimates of brown bear home range size sigma in meters.
}
\examples{
tab1()
}
