% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{tab2}
\alias{tab2}
\title{Table 2
Each row in the table lists the number (and percentages) of individuals
located to either region during that particular survey. From the Region A
survey, two individuals living near the coast were placed in the sea (Bothnian bay)
and Region D has some exchange with the Norwegian population.}
\usage{
tab2()
}
\value{
A data frame.
}
\description{
Table 2
Each row in the table lists the number (and percentages) of individuals
located to either region during that particular survey. From the Region A
survey, two individuals living near the coast were placed in the sea (Bothnian bay)
and Region D has some exchange with the Norwegian population.
}
\examples{
tab2()
}
