% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{tabS1}
\alias{tabS1}
\title{Table S1
Placement of individual activity centers with Negative binomial P(n)}
\usage{
tabS1()
}
\value{
A data frame.
}
\description{
Table S1
Placement of individual activity centers with Negative binomial P(n)
}
\examples{
tabS1()
}
