% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{tabS2}
\alias{tabS2}
\title{Table S2
Placement of individual activity centers with Poisson P(n)}
\usage{
tabS2()
}
\value{
A data frame.
}
\description{
Table S2
Placement of individual activity centers with Poisson P(n)
}
\examples{
tabS2()

}
