% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{tabS3}
\alias{tabS3}
\title{Table S3
Number of individuals placed in neighbouring region using Likelihood based estimate of μ and Threshold.
Negative binomial P(n)}
\usage{
tabS3()
}
\value{
A data frame.
}
\description{
Table S3
Number of individuals placed in neighbouring region using Likelihood based estimate of μ and Threshold.
Negative binomial P(n)
}
\examples{
tabS3()

}
