% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{tabS4}
\alias{tabS4}
\title{Table S4
Number of individuals placed in neighbouring region using Likelihood based estimate of μ and Threshold.
Poisson P(n)}
\usage{
tabS4()
}
\value{
A data frame.
}
\description{
Table S4
Number of individuals placed in neighbouring region using Likelihood based estimate of μ and Threshold.
Poisson P(n)
}
\examples{
tabS4()

}
