% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{threshold_NB}
\alias{threshold_NB}
\title{Calculate the threshold for the Negative Binomial distribution}
\usage{
threshold_NB(k, sigma, lambda, theta)
}
\arguments{
\item{k}{number of individuals}

\item{sigma}{standard deviation of home range distribution}

\item{lambda}{mean of the Negative Binomial distribution}

\item{theta}{dispersion parameter of the Negative Binomial distribution}
}
\value{
threshold for the Negative Binomial distribution
}
\description{
Calculate the threshold for the Negative Binomial distribution
}
