% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{threshold_Po}
\alias{threshold_Po}
\title{Calculate the threshold for the Poisson distribution}
\usage{
threshold_Po(k, sigma, lambda)
}
\arguments{
\item{k}{number of individuals}

\item{sigma}{standard deviation of home range distribution}

\item{lambda}{mean of the Poisson distribution}
}
\value{
threshold for the Poisson distribution
}
\description{
Calculate the threshold for the Poisson distribution
}
